requirements.txt
langgraph>=0.2.28
langchain>=0.3.0
langchain-google-genai>=4.0.0
google-genai>=1.7.0
python-dotenv==1.0.1
pydantic>=2.7.4
